refactor: add vendor comment form and secondary fields components, restructure title and description handling
