+++ apps/app/src/app/[locale]/(app)/(dashboard)/[orgId]/vendors/[vendorId]/page.tsx
@@ -15,8 +15,8 @@ import { Card, CardContent, CardHeader, CardTitle } from "@bubba/ui/card";
 import type { Metadata } from "next";
 import { setStaticParamsLocale } from "next-international/server";
 import { redirect } from "next/navigation";
-import { TitleAndDescription } from "./components/title-and-description";
-import { SecondaryFields } from "./components/secondary-fields";
+import { TitleAndDescription } from "./components/title-and-description/title-and-description";
+import { SecondaryFields } from "./components/secondary-fields/secondary-fields";
 
 interface PageProps {
   searchParams: Promise<{
